Solid Version (#20)
The solid version has parody with the Qwik version, but is less strict about trying to do everything on the server.
This makes it so that we can subscribe to supabase changes, but means the bundle size is larger.
Initial values are loaded on the server, and tanstack query hydration is used to smoothly transfer that state. 
Also, there's optimistic updates, which is cool.